requirements.txt
# UFormer Star Removal Dependencies

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0

# Progress & Logging  
tqdm>=4.64.0
tensorboard>=2.12.0
pyyaml>=6.0

# Metrics & Evaluation
scikit-image>=0.20.0  # For SSIM, PSNR

# GUI (for testing interface)
tkinter  # Usually comes with Python

# Optional: Jupyter for notebooks
# jupyter>=1.0.0
# matplotlib>=3.5.0

# Development Tools (optional)
# black>=22.0.0
# flake8>=5.0.0
